Import and apply heart rate data to a spiroergometric data file

hr_add() imports an additional file containing heart rate data and adds it to an existing spiroergometric data file.

hr_add(data, hr_file, hr_offset = 0)

hr_import(hr_file)

hr_apply(data, hr_data, hr_offset = 0)

Arguments

data
A data.frame containing spiroergometric data interpolated to every second.

hr_file
The name of a *tcx file which contains additional heart rate data. The file will be found by partial matching of the regular expression within the working directory (and one level above). Alternatively, hr_file can be an absolute or relative path.

hr_offset
An integer, corresponding to the temporal offset of the heart-rate file. By default the start of the heart rate measurement is linked to the first load in the test protocol (i.e. the end of the pre-measure).

hr_data
A data.frame containing the imported heart rate data.

Value

A data.frame containing the spiroergometric and heart rate data.

Details

Heart rate data will be imported from a .tcx file with hr_import(). After interpolating the data to full seconds, it is then applied to a given data set by hr_apply().