Manually generate a testing protocol for spiroergometry files

spiro_protocol() manually generates a testing protocol which can be applied to spiroergometry files.

spiro_protocol(
  step.start,
  step.increment,
  step.duration,
  rest.duration,
  pre.duration,
  wu.duration,
  wu.load,
  step.count,
  rest.initial,
  testtype = NULL
)

spiro_protocol_clt(step.start, step.duration = 300, ...)

spiro_protocol_rmp(
  step.count,
  wu.load = 2.8,
  step.increment = 0.15,
  step.duration = 30,
  ...
)

spiro_protocol_gxt(
  step.count,
  step.start = 2,
  step.increment = 0.4,
  step.duration = 300,
  ...
)

Arguments

step.start
A numeric value giving the load of the first (non warm-up) step.

step.increment
A numeric value, by which load increases with every new step.

step.duration
An integer giving the duration of each step in seconds.

rest.duration
An integer giving the duration of the rest between the steps in seconds.

pre.duration
An integer giving the duration of the initial pre-measurement without load in seconds.

wu.duration
An integer giving the duration of the warm-up step in seconds.

wu.load
A numeric value giving the load of the warm-up step.

step.count
A numeric value indicating the number of steps. A non-integer value corresponds to a not finished last step.

rest.initial
A logical, whether there should be a initial rest between warm-up and first step.

testtype
A character value, usually "ramp", "increment" or "constant". Per default the testtype is guessed from the test protocol characteristics.

...
Passing of the above arguments in the shorthand functions.

Value

A data.frame containing the test characteristics, which can be applied to a given (interpolated) file by apply_protocol.

Details

This function provides a manual interface for generating testing protocols in exercise science. The shorthands _clt, _gxt and _rmp provide wrappers with pre-defined settings for constant load, graded exercise (i.e. stepwise incremental) and ramp tests.

To automatically determine the protocol from a data file use guess_protocol instead.

Functions

spiro_protocol_clt: Interface for constant load test protocols. Per default, six steps of five minutes with 30 seconds rest in between are applied. Following a one minute pre-measure, the warm-up is of the same length and half the load of the steps.

spiro_protocol_rmp: Interface for ramp test protocols. Per default, 30s-steps without rest in between are applied. Following a one-minute pre measure , the warm-up has a duration of two minutes and is directly followed by the first step.

spiro_protocol_gxt: Interface for graded exercise/incremental step test protocols. Per default, there is only a one-minute pre-measure with no warm-up. The default rest between steps is set to 30 seconds.

Examples

spiro_protocol(step.start = 150, step.increment = 30, step.duration = 180,
               rest.duration = 0, pre.duration = 60, wu.duration = 0,
               wu.load = 0, step.count = 7, rest.initial = FALSE)

#>   step.start step.increment step.duration rest.duration pre.duration
#> 1        150             30           180             0           60
#>   wu.duration wu.load step.count rest.initial  testtype
#> 1           0       0          7        FALSE increment

# which can be simplified:
spiro_protocol_gxt(step.count = 7, step.start = 150, step.increment = 30,
                   step.duration = 180, rest.duration = 30)

#>   step.start step.increment step.duration rest.duration pre.duration
#> 1        150             30           180            30           60
#>   wu.duration wu.load step.count rest.initial  testtype
#> 1           0       0          7        FALSE increment

spiro_protocol_rmp(step.count = 19, wu.load = 3.0, step.increment = 0.2)

#>   step.start step.increment step.duration rest.duration pre.duration
#> 1        3.2            0.2            30             0           60
#>   wu.duration wu.load step.count rest.initial testtype
#> 1         120       3         19        FALSE     ramp

spiro_protocol_clt(step.start = 4.5)

#>   step.start step.increment step.duration rest.duration pre.duration
#> 1        4.5              0           300            30           60
#>   wu.duration wu.load step.count rest.initial testtype
#> 1         300    2.25          6         TRUE constant