Smooth data with a (zero-phase) Butterworth filter

Internal function for spiro_smooth.

Usage

bw_filter(x, n = 3, W = 0.04, zero_lag = TRUE)

Arguments

x

A numeric vector on which the digital filter should be applied

n

Order of the Butterworth filter, defaults to 3

W

Low-pass cut-off frequency of the filter, defaults to 0.04

zero_lag

Whether a zero phase (forwards-backwards) filter should be applied.

Value

A numeric vector of the same length as x.

Details

Digital filtering might be a preferable processing strategy for smoothing data from gas exchange measures when compared to moving averages. Robergs et al. (2010) proposes a third order Butterworth filter with a low-pass cut-off frequency of 0.04 for filtering VO2 data.

It should be noted that Butterworth filter comprise a time lag. A method to create a data series with zero lag is to subsequently apply two Butterworth filters in forward and reverse direction (forward-backwards filtering). While this procedure removes any time lag it changes the magnitude of the filtering response, i.e. the resulting filter has not the same properties (order and cut-off frequency) as a single filter has.

Examples

# Get VO2 data from example file
vo2_vector <- spiro(spiro_example("zan_gxt"))$VO2

bw_filter(vo2_vector)
#>    [1]  501.8269  502.6393  504.2423  506.5931  509.6280  513.2645  517.4021
#>    [8]  521.9252  526.7046  531.6010  536.4678  541.1546  545.5111  549.3910
#>   [15]  552.6563  555.1817  556.8583  557.5975  557.3338  556.0264  553.6605
#>   [22]  550.2461  545.8173  540.4291  534.1547  527.0822  519.3116  510.9522
#>   [29]  502.1201  492.9362  483.5242  474.0088  464.5150  455.1669  446.0879
#>   [36]  437.4003  429.2255  421.6819  414.8844  408.9412  403.9527  400.0086
#>   [43]  397.1874  395.5547  395.1637  396.0555  398.2604  401.7994  406.6862
#>   [50]  412.9294  420.5344  429.5049  439.8448  451.5588  464.6535  479.1382
#>   [57]  495.0241  512.3247  531.0541  551.2255  572.8491  595.9303  620.4674
#>   [64]  646.4501  673.8592  702.6664  732.8350  764.3206  797.0712  831.0283
#>   [71]  866.1267  902.2961  939.4617  977.5453 1016.4664 1056.1413 1096.4828
#>   [78] 1137.3989 1178.7919 1220.5579 1262.5860 1304.7562 1346.9382 1388.9906
#>   [85] 1430.7615 1472.0905 1512.8104 1552.7494 1591.7343 1629.5949 1666.1684
#>   [92] 1701.3037 1734.8635 1766.7261 1796.7861 1824.9538 1851.1550 1875.3288
#>   [99] 1897.4271 1917.4129 1935.2612 1950.9597 1964.5112 1975.9356 1985.2715
#>  [106] 1992.5768 1997.9300 2001.4304 2003.2000 2003.3822 2002.1389 1999.6457
#>  [113] 1996.0872 1991.6516 1986.5274 1980.8992 1974.9451 1968.8340 1962.7229
#>  [120] 1956.7557 1951.0599 1945.7456 1940.9025 1936.5997 1932.8839 1929.7799
#>  [127] 1927.2902 1925.3963 1924.0598 1923.2262 1922.8282 1922.7907 1923.0344
#>  [134] 1923.4807 1924.0553 1924.6909 1925.3293 1925.9228 1926.4362 1926.8494
#>  [141] 1927.1613 1927.3912 1927.5787 1927.7803 1928.0642 1928.5048 1929.1766
#>  [148] 1930.1489 1931.4808 1933.2160 1935.3788 1937.9713 1940.9721 1944.3381
#>  [155] 1948.0078 1951.9067 1955.9526 1960.0608 1964.1475 1968.1334 1971.9451
#>  [162] 1975.5173 1978.7938 1981.7289 1984.2892 1986.4552 1988.2224 1989.6011
#>  [169] 1990.6134 1991.2899 1991.6667 1991.7821 1991.6746 1991.3802 1990.9301
#>  [176] 1990.3498 1989.6584 1988.8692 1987.9903 1987.0260 1985.9768 1984.8417
#>  [183] 1983.6189 1982.3088 1980.9168 1979.4548 1977.9426 1976.4070 1974.8789
#>  [190] 1973.3917 1971.9781 1970.6688 1969.4893 1968.4595 1967.5916 1966.8908
#>  [197] 1966.3567 1965.9859 1965.7759 1965.7275 1965.8467 1966.1448 1966.6371
#>  [204] 1967.3398 1968.2660 1969.4200 1970.7941 1972.3658 1974.0979 1975.9407
#>  [211] 1977.8363 1979.7224 1981.5359 1983.2166 1984.7091 1985.9655 1986.9460
#>  [218] 1987.6193 1987.9616 1987.9556 1987.5889 1986.8539 1985.7464 1984.2659
#>  [225] 1982.4156 1980.2021 1977.6346 1974.7240 1971.4823 1967.9226 1964.0594
#>  [232] 1959.9098 1955.4946 1950.8391 1945.9730 1940.9303 1935.7488 1930.4693
#>  [239] 1925.1355 1919.7929 1914.4871 1909.2623 1904.1586 1899.2112 1894.4498
#>  [246] 1889.8990 1885.5793 1881.5082 1877.7009 1874.1715 1870.9332 1867.9996
#>  [253] 1865.3849 1863.1044 1861.1745 1859.6121 1858.4337 1857.6534 1857.2815
#>  [260] 1857.3228 1857.7771 1858.6400 1859.9039 1861.5593 1863.5943 1865.9955
#>  [267] 1868.7470 1871.8305 1875.2236 1878.8987 1882.8223 1886.9542 1891.2471
#>  [274] 1895.6480 1900.0985 1904.5371 1908.9011 1913.1277 1917.1555 1920.9245
#>  [281] 1924.3762 1927.4540 1930.1027 1932.2689 1933.9008 1934.9497 1935.3718
#>  [288] 1935.1308 1934.2004 1932.5659 1930.2250 1927.1879 1923.4758 1919.1197
#>  [295] 1914.1585 1908.6388 1902.6128 1896.1385 1889.2781 1882.0980 1874.6676
#>  [302] 1867.0597 1859.3502 1851.6178 1843.9435 1836.4097 1829.0997 1822.0966
#>  [309] 1815.4823 1809.3353 1803.7287 1798.7283 1794.3912 1790.7655 1787.8908
#>  [316] 1785.7987 1784.5136 1784.0519 1784.4213 1785.6202 1787.6362 1790.4469
#>  [323] 1794.0195 1798.3112 1803.2695 1808.8317 1814.9257 1821.4696 1828.3735
#>  [330] 1835.5408 1842.8704 1850.2584 1857.5993 1864.7869 1871.7152 1878.2789
#>  [337] 1884.3741 1889.8993 1894.7565 1898.8512 1902.0931 1904.3943 1905.6694
#>  [344] 1905.8349 1904.8092 1902.5136 1898.8719 1893.8115 1887.2633 1879.1635
#>  [351] 1869.4548 1858.0888 1845.0283 1830.2500 1813.7473 1795.5331 1775.6413
#>  [358] 1754.1278 1731.0713 1706.5733 1680.7584 1653.7740 1625.7895 1596.9942
#>  [365] 1567.5940 1537.8084 1507.8652 1477.9973 1448.4380 1419.4194 1391.1697
#>  [372] 1363.9119 1337.8615 1313.2248 1290.1963 1268.9570 1249.6730 1232.4943
#>  [379] 1217.5539 1204.9675 1194.8322 1187.2255 1182.2039 1179.8004 1180.0232
#>  [386] 1182.8530 1188.2416 1196.1101 1206.3487 1218.8167 1233.3429 1249.7270
#>  [393] 1267.7409 1287.1305 1307.6197 1328.9150 1350.7131 1372.7093 1394.6074
#>  [400] 1416.1288 1437.0219 1457.0694 1476.0937 1493.9600 1510.5773 1525.8972
#>  [407] 1539.9116 1552.6481 1564.1664 1574.5530 1583.9160 1592.3809 1600.0859
#>  [414] 1607.1791 1613.8149 1620.1516 1626.3481 1632.5628 1638.9516 1645.6667
#>  [421] 1652.8557 1660.6596 1669.2100 1678.6266 1689.0128 1700.4532 1713.0106
#>  [428] 1726.7247 1741.6110 1757.6610 1774.8423 1793.1003 1812.3592 1832.5238
#>  [435] 1853.4815 1875.1037 1897.2480 1919.7597 1942.4744 1965.2214 1987.8280
#>  [442] 2010.1240 2031.9456 2053.1391 2073.5630 2093.0884 2111.6000 2128.9953
#>  [449] 2145.1842 2160.0879 2173.6379 2185.7758 2196.4532 2205.6316 2213.2830
#>  [456] 2219.3902 2223.9471 2226.9604 2228.4505 2228.4540 2227.0253 2224.2373
#>  [463] 2220.1812 2214.9649 2208.7106 2201.5521 2193.6320 2185.0987 2176.1042
#>  [470] 2166.8011 2157.3400 2147.8664 2138.5180 2129.4217 2120.6912 2112.4249
#>  [477] 2104.7039 2097.5920 2091.1369 2085.3714 2080.3150 2075.9760 2072.3531
#>  [484] 2069.4370 2067.2118 2065.6552 2064.7395 2064.4325 2064.6985 2065.4990
#>  [491] 2066.7930 2068.5378 2070.6901 2073.2059 2076.0416 2079.1537 2082.4994
#>  [498] 2086.0375 2089.7292 2093.5388 2097.4336 2101.3828 2105.3580 2109.3338
#>  [505] 2113.2891 2117.2086 2121.0838 2124.9129 2128.7008 2132.4590 2136.2061
#>  [512] 2139.9671 2143.7741 2147.6650 2151.6824 2155.8722 2160.2827 2164.9626
#>  [519] 2169.9590 2175.3144 2181.0617 2187.2183 2193.7798 2200.7140 2207.9574
#>  [526] 2215.4141 2222.9585 2230.4416 2237.6998 2244.5654 2250.8774 2256.4896
#>  [533] 2261.2768 2265.1377 2267.9964 2269.8010 2270.5222 2270.1487 2268.6839
#>  [540] 2266.1408 2262.5385 2257.8984 2252.2429 2245.5953 2237.9813 2229.4320
#>  [547] 2219.9863 2209.6937 2198.6158 2186.8286 2174.4226 2161.5018 2148.1794
#>  [554] 2134.5728 2120.7979 2106.9641 2093.1721 2079.5137 2066.0742 2052.9353
#>  [561] 2040.1774 2027.8817 2016.1318 2005.0155 1994.6261 1985.0629 1976.4288
#>  [568] 1968.8268 1962.3530 1957.0897 1953.0997 1950.4234 1949.0793 1949.0647
#>  [575] 1950.3585 1952.9222 1956.7026 1961.6334 1967.6386 1974.6350 1982.5342
#>  [582] 1991.2449 2000.6737 2010.7263 2021.3085 2032.3273 2043.6918 2055.3147
#>  [589] 2067.1132 2079.0093 2090.9294 2102.8025 2114.5583 2126.1250 2137.4281
#>  [596] 2148.3889 2158.9251 2168.9514 2178.3813 2187.1299 2195.1157 2202.2630
#>  [603] 2208.5019 2213.7690 2218.0058 2221.1580 2223.1751 2224.0114 2223.6288
#>  [610] 2222.0008 2219.1152 2214.9775 2209.6115 2203.0607 2195.3872 2186.6712
#>  [617] 2177.0080 2166.5070 2155.2892 2143.4850 2131.2317 2118.6701 2105.9419
#>  [624] 2093.1872 2080.5438 2068.1471 2056.1303 2044.6250 2033.7600 2023.6592
#>  [631] 2014.4386 2006.2026 1999.0413 1993.0272 1988.2131 1984.6297 1982.2843
#>  [638] 1981.1606 1981.2201 1982.4050 1984.6418 1987.8449 1991.9193 1996.7645
#>  [645] 2002.2782 2008.3603 2014.9160 2021.8577 2029.1058 2036.5890 2044.2439
#>  [652] 2052.0139 2059.8483 2067.7005 2075.5271 2083.2869 2090.9414 2098.4549
#>  [659] 2105.7957 2112.9359 2119.8521 2126.5242 2132.9343 2139.0658 2144.9009
#>  [666] 2150.4180 2155.5889 2160.3756 2164.7283 2168.5855 2171.8738 2174.5091
#>  [673] 2176.3963 2177.4302 2177.4940 2176.4594 2174.1857 2170.5210 2165.3043
#>  [680] 2158.3685 2149.5456 2138.6725 2125.5981 2110.1899 2092.3412 2071.9774
#>  [687] 2049.0628 2023.6065 1995.6666 1965.3546 1932.8368 1898.3351 1862.1242
#>  [694] 1824.5269 1785.9054 1746.6525 1707.1797 1667.9066 1629.2515 1591.6222
#>  [701] 1555.4094 1520.9798 1488.6709 1458.7854 1431.5872 1407.2976 1386.0935
#>  [708] 1368.1061 1353.4213 1342.0808 1334.0843 1329.3920 1327.9273 1329.5797
#>  [715] 1334.2074 1341.6410 1351.6862 1364.1282 1378.7352 1395.2613 1413.4488
#>  [722] 1433.0293 1453.7243 1475.2472 1497.3058 1519.6073 1541.8643 1563.8032
#>  [729] 1585.1710 1605.7432 1625.3288 1643.7737 1660.9632 1676.8231 1691.3201
#>  [736] 1704.4613 1716.2929 1726.8964 1736.3844 1744.8955 1752.5886 1759.6374
#>  [743] 1766.2238 1772.5327 1778.7458 1785.0365 1791.5653 1798.4762 1805.8929
#>  [750] 1813.9176 1822.6301 1832.0890 1842.3332 1853.3831 1865.2419 1877.8966
#>  [757] 1891.3193 1905.4693 1920.2946 1935.7340 1951.7191 1968.1762 1985.0290
#>  [764] 2002.2010 2019.6193 2037.2163 2054.9314 2072.7110 2090.5086 2108.2829
#>  [771] 2125.9964 2143.6147 2161.1048 2178.4343 2195.5712 2212.4823 2229.1327
#>  [778] 2245.4846 2261.4961 2277.1212 2292.3094 2307.0067 2321.1567 2334.7026
#>  [785] 2347.5883 2359.7604 2371.1685 2381.7669 2391.5162 2400.3851 2408.3523
#>  [792] 2415.4085 2421.5570 2426.8143 2431.2105 2434.7881 2437.6021 2439.7178
#>  [799] 2441.2094 2442.1573 2442.6456 2442.7591 2442.5804 2442.1877 2441.6527
#>  [806] 2441.0398 2440.4050 2439.7952 2439.2480 2438.7914 2438.4437 2438.2140
#>  [813] 2438.1022 2438.0997 2438.1917 2438.3596 2438.5836 2438.8445 2439.1245
#>  [820] 2439.4065 2439.6735 2439.9064 2440.0841 2440.1823 2440.1749 2440.0355
#>  [827] 2439.7388 2439.2626 2438.5899 2437.7098 2436.6197 2435.3247 2433.8379
#>  [834] 2432.1783 2430.3698 2428.4397 2426.4185 2424.3378 2422.2300 2420.1261
#>  [841] 2418.0556 2416.0454 2414.1204 2412.3035 2410.6163 2409.0793 2407.7114
#>  [848] 2406.5292 2405.5465 2404.7730 2404.2148 2403.8747 2403.7535 2403.8516
#>  [855] 2404.1694 2404.7076 2405.4664 2406.4436 2407.6331 2409.0224 2410.5911
#>  [862] 2412.3095 2414.1386 2416.0315 2417.9362 2419.7993 2421.5698 2423.2033
#>  [869] 2424.6642 2425.9286 2426.9845 2427.8333 2428.4891 2428.9794 2429.3438
#>  [876] 2429.6333 2429.9076 2430.2326 2430.6762 2431.3051 2432.1823 2433.3656
#>  [883] 2434.9070 2436.8509 2439.2329 2442.0775 2445.3960 2449.1842 2453.4210
#>  [890] 2458.0674 2463.0670 2468.3475 2473.8232 2479.3983 2484.9707 2490.4360
#>  [897] 2495.6914 2500.6395 2505.1914 2509.2694 2512.8081 2515.7550 2518.0706
#>  [904] 2519.7279 2520.7125 2521.0228 2520.6691 2519.6727 2518.0646 2515.8842
#>  [911] 2513.1775 2509.9949 2506.3890 2502.4122 2498.1140 2493.5388 2488.7241
#>  [918] 2483.6988 2478.4837 2473.0923 2467.5336 2461.8151 2455.9460 2449.9391
#>  [925] 2443.8130 2437.5931 2431.3120 2425.0110 2418.7397 2412.5573 2406.5316
#>  [932] 2400.7385 2395.2600 2390.1815 2385.5890 2381.5654 2378.1876 2375.5227
#>  [939] 2373.6257 2372.5369 2372.2798 2372.8587 2374.2567 2376.4337 2379.3252
#>  [946] 2382.8426 2386.8739 2391.2872 2395.9352 2400.6624 2405.3124 2409.7354
#>  [953] 2413.7951 2417.3733 2420.3738 2422.7241 2424.3757 2425.3029 2425.5003
#>  [960] 2424.9792 2423.7651 2421.8942 2419.4111 2416.3669 2412.8167 2408.8186
#>  [967] 2404.4332 2399.7243 2394.7587 2389.6062 2384.3391 2379.0312 2373.7565
#>  [974] 2368.5891 2363.6020 2358.8678 2354.4591 2350.4485 2346.9082 2343.9092
#>  [981] 2341.5197 2339.8037 2338.8193 2338.6169 2339.2367 2340.7079 2343.0459
#>  [988] 2346.2517 2350.3109 2355.1924 2360.8479 2367.2101 2374.1931 2381.6933
#>  [995] 2389.5925 2397.7609 2406.0608 2414.3485 2422.4752 2430.2877 2437.6269
#> [1002] 2444.3263 2450.2102 2455.0922 2458.7757 2461.0540 2461.7129 2460.5329
#> [1009] 2457.2922 2451.7710 2443.7562 2433.0477 2419.4649 2402.8527 2383.0876
#> [1016] 2360.0830 2333.7947 2304.2254 2271.4292 2235.5151 2196.6491 2155.0555
#> [1023] 2111.0158 2064.8672 2016.9974 1967.8395 1917.8642 1867.5707 1817.4772
#> [1030] 1768.1106 1719.9962 1673.6475 1629.5570 1588.1869 1549.9621 1515.2639
#> [1037] 1484.4246 1457.7243 1435.3880 1417.5836 1404.4214 1395.9539 1392.1763
#> [1044] 1393.0272 1398.3892 1408.0901 1421.9035 1439.5510 1460.7048 1484.9931
#> [1051] 1512.0056 1541.3012 1572.4167 1604.8761 1638.2008 1671.9208 1705.5856
#> [1058] 1738.7756 1771.1127 1802.2692 1831.9748 1860.0210 1886.2632 1910.6195
#> [1065] 1933.0688 1953.6460 1972.4356 1989.5653 2005.1979 2019.5242 2032.7546
#> [1072] 2045.1106 2056.8165 2068.0910 2079.1398 2090.1493 2101.2821 2112.6739
#> [1079] 2124.4341 2136.6479 2149.3788 2162.6723 2176.5569 2191.0460 2206.1379
#> [1086] 2221.8161 2238.0507 2254.7993 2272.0088 2289.6172 2307.5554 2325.7487
#> [1093] 2344.1184 2362.5846 2381.0679 2399.4930 2417.7903 2435.8987 2453.7659
#> [1100] 2471.3486 2488.6120 2505.5280 2522.0745 2538.2339 2553.9914 2569.3335
#> [1107] 2584.2470 2598.7191 2612.7374 2626.2910 2639.3709 2651.9696 2664.0816
#> [1114] 2675.7030 2686.8319 2697.4671 2707.6080 2717.2534 2726.4016 2735.0502
#> [1121] 2743.1969 2750.8387 2757.9715 2764.5894 2770.6835 2776.2415 2781.2482
#> [1128] 2785.6874 2789.5432 2792.8022 2795.4549 2797.4969 2798.9281 2799.7523
#> [1135] 2799.9767 2799.6112 2798.6699 2797.1717 2795.1407 2792.6073 2789.6072
#> [1142] 2786.1802 2782.3695 2778.2201 2773.7792 2769.0950 2764.2163 2759.1914
#> [1149] 2754.0677 2748.8924 2743.7132 2738.5805 2733.5487 2728.6776 2724.0334
#> [1156] 2719.6884 2715.7206 2712.2120 2709.2451 2706.8993 2705.2462 2704.3459
#> [1163] 2704.2436 2704.9681 2706.5304 2708.9240 2712.1239 2716.0864 2720.7484
#> [1170] 2726.0270 2731.8200 2738.0072 2744.4536 2751.0138 2757.5377 2763.8770
#> [1177] 2769.8910 2775.4508 2780.4428 2784.7712 2788.3595 2791.1505 2793.1064
#> [1184] 2794.2070 2794.4490 2793.8435 2792.4158 2790.2044 2787.2607 2783.6495
#> [1191] 2779.4480 2774.7450 2769.6379 2764.2288 2758.6205 2752.9119 2747.1942
#> [1198] 2741.5480 2736.0411 2730.7273 2725.6465 2720.8261 2716.2815 2712.0187
#> [1205] 2708.0347 2704.3185 2700.8523 2697.6109 2694.5631 2691.6721 2688.8962
#> [1212] 2686.1906 2683.5091 2680.8057 2678.0367 2675.1615 2672.1446 2668.9563
#> [1219] 2665.5735 2661.9798 2658.1662 2654.1305 2649.8774 2645.4179 2640.7697
#> [1226] 2635.9578 2631.0163 2625.9894 2620.9324 2615.9119 2611.0044 2606.2943
#> [1233] 2601.8712 2597.8260 2594.2481 2591.2214 2588.8213 2587.1118 2586.1448
#> [1240] 2585.9597 2586.5840 2588.0339 2590.3146 2593.4214 2597.3391 2602.0431
#> [1247] 2607.4998 2613.6681 2620.5013 2627.9490 2635.9583 2644.4752 2653.4449
#> [1254] 2662.8124 2672.5221 2682.5170 2692.7377 2703.1199 2713.5932 2724.0795
#> [1261] 2734.4913 2744.7320 2754.6960 2764.2708 2773.3403 2781.7874 2789.4976
#> [1268] 2796.3617 2802.2794 2807.1636 2810.9456 2813.5794 2815.0459 2815.3542
#> [1275] 2814.5426 2812.6770 2809.8490 2806.1727 2801.7815 2796.8243 2791.4618
#> [1282] 2785.8624 2780.1976 2774.6359 2769.3373 2764.4463 2760.0857 2756.3510
#> [1289] 2753.3060 2750.9792 2749.3645 2748.4249 2748.0983 2748.3059 2748.9586
#> [1296] 2749.9629 2751.2253 2752.6567 2754.1758 2755.7127 2757.2104 2758.6248
#> [1303] 2759.9244 2761.0885 2762.1062 2762.9749 2763.6990 2764.2893 2764.7614
#> [1310] 2765.1355 2765.4349 2765.6840 2765.9063 2766.1224 2766.3482 2766.5944
#> [1317] 2766.8658 2767.1622 2767.4785 2767.8048 2768.1270 2768.4263 2768.6800
#> [1324] 2768.8605 2768.9351 2768.8649 2768.6033 2768.0940 2767.2684 2766.0433
#> [1331] 2764.3191 2761.9785 2758.8850 2754.8821 2749.7912 2743.4113 2735.5197
#> [1338] 2725.8746 2714.2210 2700.2980 2683.8486 2664.6304 2642.4275 2617.0612
#> [1345] 2588.3996 2556.3658 2520.9437 2482.1828 2440.2005 2395.1829 2347.3824
#> [1352] 2297.1165 2244.7633 2190.7584 2135.5878 2079.7809 2023.9007 1968.5329
#> [1359] 1914.2735 1861.7158 1811.4381 1763.9915 1719.8901 1679.6017 1643.5415
#> [1366] 1612.0661 1585.4694 1563.9788 1547.7526 1536.8784 1531.3724 1531.1797
#> [1373] 1536.1758 1546.1681 1560.8994 1580.0520 1603.2532 1630.0826 1660.0801
#> [1380] 1692.7559 1727.6008 1764.0972 1801.7301 1839.9973 1878.4193 1916.5493
#> [1387] 1953.9825 1990.3646 2025.3987 2058.8505 2090.5516 2120.4003 2148.3604
#> [1394] 2174.4573 2198.7719 2221.4332 2242.6083 2262.4932 2281.3019 2299.2559
#> [1401] 2316.5765 2333.4784 2350.1664 2366.8329 2383.6552 2400.7924 2418.3823
#> [1408] 2436.5368 2455.3393 2474.8424 2495.0660 2515.9979 2537.5941 2559.7802
#> [1415] 2582.4534 2605.4836 2628.7162 2651.9766 2675.0775 2697.8277 2720.0398
#> [1422] 2741.5370 2762.1578 2781.7602 2800.2245 2817.4567 2833.3891 2847.9814
#> [1429] 2861.2188 2873.1101 2883.6859 2892.9968 2901.1116 2908.1161 2914.1106
#> [1436] 2919.2070 2923.5249 2927.1870 2930.3154 2933.0278 2935.4350 2937.6386
#> [1443] 2939.7282 2941.7797 2943.8538 2945.9958 2948.2367 2950.5948 2953.0781
#> [1450] 2955.6873 2958.4176 2961.2595 2964.1996 2967.2203 2970.2999 2973.4127
#> [1457] 2976.5285 2979.6126 2982.6266 2985.5291 2988.2777 2990.8300 2993.1451
#> [1464] 2995.1849 2996.9157 2998.3093 2999.3449 3000.0098 3000.3012 3000.2256
#> [1471] 2999.7991 2999.0478 2998.0067 2996.7210 2995.2453 2993.6438 2991.9893
#> [1478] 2990.3622 2988.8479 2987.5338 2986.5061 2985.8463 2985.6273 2985.9108
#> [1485] 2986.7451 2988.1633 2990.1844 2992.8136 2996.0436 2999.8557 3004.2201
#> [1492] 3009.0970 3014.4381 3020.1888 3026.2901 3032.6812 3039.3017 3046.0926
#> [1499] 3052.9976 3059.9638 3066.9409 3073.8820 3080.7426 3087.4819 3094.0621
#> [1506] 3100.4499 3106.6161 3112.5370 3118.1933 3123.5694 3128.6512 3133.4243
#> [1513] 3137.8710 3141.9691 3145.6898 3148.9962 3151.8440 3154.1840 3155.9661
#> [1520] 3157.1442 3157.6799 3157.5451 3156.7232 3155.2105 3153.0164 3150.1641
#> [1527] 3146.6908 3142.6473 3138.0981 3133.1203 3127.8021 3122.2406 3116.5396
#> [1534] 3110.8066 3105.1515 3099.6835 3094.5083 3089.7243 3085.4193 3081.6668
#> [1541] 3078.5243 3076.0323 3074.2129 3073.0696 3072.5860 3072.7248 3073.4266
#> [1548] 3074.6092 3076.1691 3077.9854 3079.9256 3081.8541 3083.6401 3085.1637
#> [1555] 3086.3205 3087.0227 3087.1993 3086.7943 3085.7633 3084.0706 3081.6870
#> [1562] 3078.5898 3074.7644 3070.2072 3064.9283 3058.9535 3052.3268 3045.1115
#> [1569] 3037.3903 3029.2638 3020.8481 3012.2718 3003.6719 2995.1907 2986.9706
#> [1576] 2979.1505 2971.8609 2965.2198 2959.3291 2954.2717 2950.1102 2946.8868
#> [1583] 2944.6248 2943.3313 2943.0007 2943.6179 2945.1621 2947.6088 2950.9311
#> [1590] 2955.0994 2960.0801 2965.8337 2972.3124 2979.4591 2987.2068 2995.4788
#> [1597] 3004.1913 3013.2555 3022.5801 3032.0729 3041.6418 3051.1952 3060.6414
#> [1604] 3069.8896 3078.8501 3087.4359 3095.5650 3103.1620 3110.1601 3116.5018
#> [1611] 3122.1390 3127.0323 3131.1492 3134.4631 3136.9525 3138.6024 3139.4061
#> [1618] 3139.3685 3138.5094 3136.8646 3134.4856 3131.4383 3127.8005 3123.6594
#> [1625] 3119.1098 3114.2519 3109.1894 3104.0269 3098.8676 3093.8109 3088.9493
#> [1632] 3084.3665 3080.1351 3076.3155 3072.9552 3070.0880 3067.7341 3065.9004
#> [1639] 3064.5823 3063.7654 3063.4291 3063.5485 3064.0971 3065.0479 3066.3739
#> [1646] 3068.0477 3070.0406 3072.3208 3074.8511 3077.5863 3080.4705 3083.4353
#> [1653] 3086.3986 3089.2650 3091.9265 3094.2638 3096.1478 3097.4397 3097.9915
#> [1660] 3097.6459 3096.2368 3093.5908 3089.5302 3083.8757 3076.4497 3067.0776
#> [1667] 3055.5899 3041.8224 3025.6174 3006.8241 2985.2999 2960.9130 2933.5456
#> [1674] 2903.0998 2869.5051 2832.7268 2792.7749 2749.7118 2703.6581 2654.7970
#> [1681] 2603.3760 2549.7074 2494.1664 2437.1866 2379.2534 2320.8962 2262.6779
#> [1688] 2205.1849 2149.0142 2094.7624 2043.0128 1994.3237 1949.2181 1908.1727
#> [1695] 1871.6100 1839.8909 1813.3092 1792.0880 1776.3773 1766.2524 1761.7134
#> [1702] 1762.6858 1769.0214 1780.5012 1796.8390 1817.6868 1842.6412 1871.2507
#> [1709] 1903.0236 1937.4362 1973.9419 2011.9806 2050.9876 2090.4034 2129.6841
#> [1716] 2168.3116 2205.8060 2241.7378 2275.7387 2307.5113 2336.8348 2363.5676
#> [1723] 2387.6472 2409.0865 2427.9679 2444.4351 2458.6829 2470.9474 2481.4949
#> [1730] 2490.6126 2498.6000 2505.7610 2512.3971 2518.8015 2525.2539 2532.0156
#> [1737] 2539.3258 2547.3975 2556.4150 2566.5312 2577.8658 2590.5044 2604.4979
#> [1744] 2619.8622 2636.5791 2654.5974 2673.8356 2694.1854 2715.5158 2737.6780
#> [1751] 2760.5111 2783.8467 2807.5139 2831.3447 2855.1777 2878.8626 2902.2631
#> [1758] 2925.2594 2947.7498 2969.6512 2990.8988 3011.4453 3031.2593 3050.3253
#> [1765] 3068.6433 3086.2289 3103.1130 3119.3395 3134.9628 3150.0436 3164.6440
#> [1772] 3178.8229 3192.6314 3206.1093 3219.2823 3232.1607 3244.7402 3257.0027
#> [1779] 3268.9202 3280.4575 3291.5759 3302.2363 3312.4017 3322.0388 3331.1194
#> [1786] 3339.6209 3347.5274 3354.8300 3361.5260 3367.6181 3373.1119 3378.0140
#> [1793] 3382.3299 3386.0629 3389.2133 3391.7790 3393.7565 3395.1436 3395.9424
#> [1800] 3396.1625 3395.8240 3394.9595 3393.6146 3391.8463 3389.7208 3387.3095
#> [1807] 3384.6854 3381.9196 3379.0782 3376.2202 3373.3954 3370.6427 3367.9891
#> [1814] 3365.4488 3363.0238 3360.7054 3358.4767 3356.3162 3354.2017 3352.1139
#> [1821] 3350.0392 3347.9700 3345.9047 3343.8444 3341.7909 3339.7438 3337.6987
#> [1828] 3335.6471 3333.5760 3331.4697 3329.3109 3327.0842 3324.7780 3322.3875
#> [1835] 3319.9154 3317.3725 3314.7771 3312.1533 3309.5301 3306.9385 3304.4095
#> [1842] 3301.9712 3299.6469 3297.4527 3295.3974 3293.4826 3291.7046 3290.0566
#> [1849] 3288.5325 3287.1292 3285.8500 3284.7059 3283.7175 3282.9141 3282.3330
#> [1856] 3282.0173 3282.0131 3282.3661 3283.1194 3284.3110 3285.9738 3288.1355
#> [1863] 3290.8191 3294.0430 3297.8205 3302.1589 3307.0588 3312.5127 3318.5042
#> [1870] 3325.0079 3331.9892 3339.4045 3347.2020 3355.3216 3363.6955 3372.2480
#> [1877] 3380.8959 3389.5495 3398.1134 3406.4879 3414.5700 3422.2547 3429.4373
#> [1884] 3436.0161 3441.8963 3446.9934 3451.2372 3454.5731 3456.9645 3458.3928
#> [1891] 3458.8570 3458.3731 3456.9720 3454.6976 3451.6049 3447.7581 3443.2299
#> [1898] 3438.1011 3432.4598 3426.4020 3420.0319 3413.4617 3406.8114 3400.2076
#> [1905] 3393.7807 3387.6624 3381.9811 3376.8575 3372.4002 3368.7025 3365.8382
#> [1912] 3363.8600 3362.7973 3362.6562 3363.4208 3365.0549 3367.5057 3370.7057
#> [1919] 3374.5751 3379.0230 3383.9496 3389.2479 3394.8073 3400.5170 3406.2701
#> [1926] 3411.9675 3417.5204 3422.8525 3427.9016 3432.6197 3436.9722 3440.9367
#> [1933] 3444.5001 3447.6563 3450.4032 3452.7404 3454.6679 3456.1852 3457.2905
#> [1940] 3457.9806 3458.2512 3458.0979 3457.5169 3456.5069 3455.0698 3453.2124
#> [1947] 3450.9483 3448.2987 3445.2932 3441.9693 3438.3713 3434.5488 3430.5548
#> [1954] 3426.4435 3422.2688 3418.0831 3413.9371 3409.8797 3405.9589 3402.2224
#> [1961] 3398.7200 3395.5052 3392.6366 3390.1780 3388.1972 3386.7642 3385.9473
#> [1968] 3385.8102 3386.4081 3387.7855 3389.9744 3392.9924 3396.8407 3401.5022
#> [1975] 3406.9400 3413.0958 3419.8888 3427.2154 3434.9483 3442.9379 3451.0130
#> [1982] 3458.9833 3466.6403 3473.7596 3480.1025 3485.4200 3489.4564 3491.9545
#> [1989] 3492.6606 3491.3282 3487.7209 3481.6156 3472.8042 3461.0971 3446.3254
#> [1996] 3428.3443 3407.0356 3382.3116 3354.1173 3322.4334 3287.2785 3248.7120
#> [2003] 3206.8362 3161.7978 3113.7887 3063.0461 3009.8509 2954.5259 2897.4330
#> [2010] 2838.9705 2779.5690 2719.6867 2659.8021 2600.4062 2541.9925 2485.0483
#> [2017] 2430.0454 2377.4323 2327.6276 2281.0160 2237.9465 2198.7320 2163.6486
#> [2024] 2132.9340 2106.7854 2085.3566 2068.7540 2057.0339 2050.2002 2048.2032
#> [2031] 2050.9413 2058.2635 2069.9738 2085.8356 2105.5755 2128.8867 2155.4319
#> [2038] 2184.8452 2216.7345 2250.6839 2286.2569 2322.9996 2360.4450 2398.1187
#> [2045] 2435.5469 2472.2663 2507.8369 2541.8548 2573.9653 2603.8732 2631.3497
#> [2052] 2656.2369 2678.4482 2697.9677 2714.8466 2729.1976 2741.1887 2751.0345
#> [2059] 2758.9879 2765.3307 2770.3648 2774.4031 2777.7615 2780.7513 2783.6723
#> [2066] 2786.8067 2790.4149 2794.7313 2799.9612 2806.2795 2813.8294 2822.7228
#> [2073] 2833.0413 2844.8381 2858.1402 2872.9508 2889.2522 2907.0080 2926.1656
#> [2080] 2946.6586 2968.4094 2991.3317 3015.3328 3040.3164 3066.1849 3092.8410
#> [2087] 3120.1892 3148.1360 3176.5891 3205.4556 3234.6390 3264.0355 3293.5313
#> [2094] 3323.0008 3352.3082 3381.3107 3409.8630 3437.8214 3465.0456 3491.4005
#> [2101] 3516.7576 3540.9958 3564.0046 3585.6856 3605.9542 3624.7403 3641.9883
#> [2108] 3657.6570 3671.7213 3684.1721 3695.0182 3704.2857 3712.0175 3718.2714
#> [2115] 3723.1180 3726.6368 3728.9139 3730.0386 3730.1032 3729.2022 3727.4328
#> [2122] 3724.8947 3721.6908 3717.9259 3713.7068 3709.1405 3704.3334 3699.3902
#> [2129] 3694.4127 3689.4987 3684.7402 3680.2215 3676.0165 3672.1859 3668.7755
#> [2136] 3665.8144 3663.3144 3661.2702 3659.6596 3658.4449 3657.5759 3656.9933
#> [2143] 3656.6326 3656.4282 3656.3171 3656.2417 3656.1519 3656.0056 3655.7688
#> [2150] 3655.4161 3654.9308 3654.3060 3653.5441 3652.6566 3651.6639 3650.5943
#> [2157] 3649.4828 3648.3696 3647.2977 3646.3107 3645.4495 3644.7498 3644.2404
#> [2164] 3643.9410 3643.8623 3644.0064 3644.3678 3644.9346 3645.6895 3646.6111
#> [2171] 3647.6750 3648.8543 3650.1213 3651.4483 3652.8092 3654.1812 3655.5463
#> [2178] 3656.8929 3658.2159 3659.5170 3660.8033 3662.0859 3663.3780 3664.6931
#> [2185] 3666.0435 3667.4383 3668.8830 3670.3786 3671.9223 3673.5071 3675.1224
#> [2192] 3676.7552 3678.3904 3680.0115 3681.5998 3683.1348 3684.5935 3685.9511
#> [2199] 3687.1816 3688.2600 3689.1633 3689.8724 3690.3729 3690.6552 3690.7149
#> [2206] 3690.5521 3690.1716 3689.5826 3688.7989 3687.8386 3686.7241 3685.4807
#> [2213] 3684.1352 3682.7139 3681.2410 3679.7374 3678.2203 3676.7035 3675.1980
#> [2220] 3673.7129 3672.2553 3670.8293 3669.4356 3668.0705 3666.7257 3665.3891
#> [2227] 3664.0458 3662.6808 3661.2810 3659.8376 3658.3465 3656.8089 3655.2312
#> [2234] 3653.6243 3652.0038 3650.3885 3648.8004 3647.2636 3645.8040 3644.4487
#> [2241] 3643.2251 3642.1600 3641.2788 3640.6042 3640.1554 3639.9480 3639.9938
#> [2248] 3640.3007 3640.8723 3641.7078 3642.8016 3644.1434 3645.7181 3647.5057
#> [2255] 3649.4830 3651.6243 3653.9039 3656.2976 3658.7844 3661.3481 3663.9786
#> [2262] 3666.6727 3669.4339 3672.2720 3675.2019 3678.2435 3681.4198 3684.7564
#> [2269] 3688.2794 3692.0131 3695.9779 3700.1882 3704.6509 3709.3641 3714.3164
#> [2276] 3719.4867 3724.8449 3730.3531 3735.9687 3741.6460 3747.3391 3753.0024
#> [2283] 3758.5922 3764.0678 3769.3931 3774.5383 3779.4810 3784.2069 3788.7092
#> [2290] 3792.9884 3797.0514 3800.9108 3804.5846 3808.0960 3811.4735 3814.7496
#> [2297] 3817.9588 3821.1361 3824.3140 3827.5216 3830.7843 3834.1246 3837.5612
#> [2304] 3841.1082 3844.7736 3848.5563 3852.4446 3856.4135 3860.4230 3864.4161
#> [2311] 3868.3188 3872.0391 3875.4676 3878.4772 3880.9228 3882.6422 3883.4573
#> [2318] 3883.1773 3881.6015 3878.5241 3873.7397 3867.0490 3858.2633 3847.2077
#> [2325] 3833.7204 3817.6524 3798.8662 3777.2370 3752.6554 3725.0312 3694.2978
#> [2332] 3660.4156 3623.3759 3583.2041 3539.9633 3493.7591 3444.7419 3393.1095
#> [2339] 3339.1079 3283.0301 3225.2150 3166.0438 3105.9367 3045.3478 2984.7597
#> [2346] 2924.6778 2865.6220 2808.1184 2752.6875 2699.8324 2650.0264 2603.7026
#> [2353] 2561.2441 2522.9783 2489.1721 2460.0300 2435.6922 2416.2351 2401.6711
#> [2360] 2391.9514 2386.9677 2386.5557 2390.4992 2398.5356 2410.3626 2425.6452
#> [2367] 2444.0228 2465.1150 2488.5268 2513.8529 2540.6816 2568.5987 2597.1920
#> [2374] 2626.0560 2654.8000 2683.0564 2710.4905 2736.8102 2761.7740 2785.1976
#> [2381] 2806.9587 2826.9992 2845.3248 2862.0008 2877.1465 2890.9269 2903.5441
#> [2388] 2915.2285 2926.2295 2936.8069 2947.2232 2957.7363 2968.5940 2980.0288
#> [2395] 2992.2537 3005.4590 3019.8097 3035.4428 3052.4651 3070.9512 3090.9425
#> [2402] 3112.4470 3135.4395 3159.8634 3185.6334 3212.6382 3240.7434 3269.7948
#> [2409] 3299.6222 3330.0427 3360.8655 3391.8956 3422.9381 3453.8022 3484.3056
#> [2416] 3514.2784 3543.5668 3572.0354 3599.5695 3626.0760 3651.4838 3675.7438
#> [2423] 3698.8268 3720.7219 3741.4334 3760.9779 3779.3814 3796.6759 3812.8971
#> [2430] 3828.0820 3842.2687 3855.4964 3867.8068 3879.2454 3889.8619 3899.7112
#> [2437] 3908.8516 3917.3440 3925.2495 3932.6276 3939.5342 3946.0209 3952.1347
#> [2444] 3957.9184 3963.4123 3968.6547 3973.6836 3978.5375 3983.2554 3987.8767
#> [2451] 3992.4399 3996.9826 4001.5406 4006.1473 4010.8320 4015.6193 4020.5277
#> [2458] 4025.5699 4030.7523 4036.0770 4041.5415 4047.1402 4052.8642 4058.7017
#> [2465] 4064.6385 4070.6594 4076.7488 4082.8920 4089.0757 4095.2882 4101.5187
#> [2472] 4107.7565 4113.9905 4120.2089 4126.3984 4132.5447 4138.6320 4144.6437
#> [2479] 4150.5622 4156.3699 4162.0487 4167.5808 4172.9480 4178.1317 4183.1129
#> [2486] 4187.8719 4192.3890 4196.6445 4200.6190 4204.2941 4207.6525 4210.6785
#> [2493] 4213.3581 4215.6790 4217.6309 4219.2060 4220.3994 4221.2099 4221.6401
#> [2500] 4221.6978 4221.3959 4220.7530 4219.7928 4218.5439 4217.0393 4215.3160
#> [2507] 4213.4140 4211.3755 4209.2442 4207.0642 4204.8789 4202.7305 4200.6585
#> [2514] 4198.6988 4196.8824 4195.2341 4193.7720 4192.5075 4191.4460 4190.5881
#> [2521] 4189.9288 4189.4579 4189.1605 4189.0179 4189.0099 4189.1166 4189.3211
#> [2528] 4189.6120 4189.9854 4190.4454 4191.0044 4191.6812 4192.4997 4193.4851
#> [2535] 4194.6621 4196.0516 4197.6693 4199.5248 4201.6206 4203.9527 4206.5106
#> [2542] 4209.2773 4212.2302 4215.3417 4218.5804 4221.9125 4225.3027 4228.7157
#> [2549] 4232.1168 4235.4731 4238.7538 4241.9306 4244.9774 4247.8703 4250.5882
#> [2556] 4253.1136 4255.4340 4257.5418 4259.4356 4261.1185 4262.5983 4263.8852
#> [2563] 4264.9913 4265.9293 4266.7121 4267.3536 4267.8690 4268.2757 4268.5946
#> [2570] 4268.8506 4269.0726 4269.2929 4269.5466 4269.8692 4270.2951 4270.8544
#> [2577] 4271.5718 4272.4648 4273.5440 4274.8126 4276.2669 4277.8964 4279.6838
#> [2584] 4281.6043 4283.6257 4285.7078 4287.8042 4289.8636 4291.8309 4293.6494
#> [2591] 4295.2610 4296.6080 4297.6343 4298.2863 4298.5136 4298.2698 4297.5127
#> [2598] 4296.2050 4294.3146 4291.8159 4288.6909 4284.9311 4280.5388 4275.5279
#> [2605] 4269.9244 4263.7663 4257.1037 4249.9983 4242.5233 4234.7618 4226.8054
#> [2612] 4218.7513 4210.6982 4202.7440 4194.9823 4187.5014 4180.3831 4173.7015
#> [2619] 4167.5223 4161.9027 4156.8911 4152.5282 4148.8478 4145.8773 4143.6393
#> [2626] 4142.1512 4141.4253 4141.4683 4142.2792 4143.8485 4146.1570 4149.1753
#> [2633] 4152.8650 4157.1791 4162.0632 4167.4562 4173.2907 4179.4937 4185.9862
#> [2640] 4192.6832 4199.4925 4206.3135 4213.0361 4219.5388 4225.6897 4231.3457
#> [2647] 4236.3540 4240.5524 4243.7696 4245.8262 4246.5354 4245.7039 4243.1337
#> [2654] 4238.6234 4231.9712 4222.9782 4211.4527 4197.2136 4180.0946 4159.9469
#> [2661] 4136.6429 4110.0794 4080.1834 4046.9171 4010.2837 3970.3317 3927.1575
#> [2668] 3880.9072 3831.7762 3780.0072 3725.8873 3669.7441 3611.9391 3552.8615
#> [2675] 3492.9204 3432.5376 3372.1426 3312.1666 3253.0393 3195.1855 3139.0225
#> [2682] 3084.9568 3033.3798 2984.6632 2939.1544 2897.1712 2858.9975 2824.8788
#> [2689] 2795.0183 2769.5730 2748.6506 2732.3077 2720.5484 2713.3253 2710.5413
#> [2696] 2712.0529 2717.6730 2727.1747 2740.2950 2756.7380 2776.1791 2798.2679
#> [2703] 2822.6325 2848.8843 2876.6248 2905.4551 2934.9879 2964.8589 2994.7379
#> [2710] 3024.3377 3053.4219 3081.8122 3109.3916 3136.1053 3161.9565 3186.9994
#> [2717] 3211.3301 3235.0776 3258.3943 3281.4480 3304.4140 3327.4693 3350.7862
#> [2724] 3374.5273 3398.8404 3423.8549 3449.6773 3476.3902 3504.0502 3532.6885
#> [2731] 3562.3115 3592.9015 3624.4183 3656.8005 3689.9678 3723.8227 3758.2535
#> [2738] 3793.1375 3828.3436 3863.7359 3899.1776 3934.5343 3969.6770 4004.4835
#> [2745] 4038.8392 4072.6361 4105.7734 4138.1559 4169.6942 4200.3050 4229.9118
#> [2752] 4258.4459 4285.8469 4312.0637 4337.0553 4360.7907 4383.2494 4404.4206
#> [2759] 4424.3030 4442.9036 4460.2364 4476.3208 4491.1791 4504.8357 4517.3161
#> [2766] 4528.6466 4538.8557 4547.9745 4556.0378 4563.0840 4569.1552 4574.2963
#> [2773] 4578.5543 4581.9783 4584.6192 4586.5308 4587.7698 4588.3966 4588.4745
#> [2780] 4588.0699 4587.2512 4586.0885 4584.6527 4583.0144 4581.2430 4579.4056
#> [2787] 4577.5659 4575.7838 4574.1142 4572.6066 4571.3033 4570.2389 4569.4397
#> [2794] 4568.9241 4568.7039 4568.7855 4569.1717 4569.8617 4570.8518 4572.1357
#> [2801] 4573.7044 4575.5470 4577.6501 4579.9986 4582.5751 4585.3608 4588.3348
#> [2808] 4591.4748 4594.7564 4598.1535 4601.6382 4605.1818 4608.7559 4612.3338
#> [2815] 4615.8924 4619.4125 4622.8794 4626.2822 4629.6137 4632.8699 4636.0492
#> [2822] 4639.1514 4642.1758 4645.1198 4647.9772 4650.7378 4653.3873 4655.9071
#> [2829] 4658.2757 4660.4689 4662.4614 4664.2281 4665.7454 4666.9930 4667.9560
#> [2836] 4668.6258 4669.0010 4669.0882 4668.9023 4668.4670 4667.8147 4666.9861
#> [2843] 4666.0293 4664.9994 4663.9565 4662.9648 4662.0899 4661.3964 4660.9456
#> [2850] 4660.7923 4660.9823 4661.5508 4662.5200 4663.8980 4665.6782 4667.8390
#> [2857] 4670.3443 4673.1462 4676.1869 4679.4028 4682.7279 4686.0967 4689.4462
#> [2864] 4692.7181 4695.8593 4698.8225 4701.5663 4704.0551 4706.2583 4708.1505
#> [2871] 4709.7106 4710.9220 4711.7719 4712.2515 4712.3552 4712.0810 4711.4306
#> [2878] 4710.4093 4709.0262 4707.2936 4705.2272 4702.8445 4700.1644 4697.2066
#> [2885] 4693.9913 4690.5387 4686.8697 4683.0058 4678.9697 4674.7858 4670.4805
#> [2892] 4666.0819 4661.6191 4657.1220 4652.6197 4648.1399 4643.7086 4639.3492
#> [2899] 4635.0818 4630.9231 4626.8858 4622.9785 4619.2056 4615.5682 4612.0647
#> [2906] 4608.6917 4605.4445 4602.3186 4599.3096 4596.4147 4593.6334 4590.9679
#> [2913] 4588.4238 4586.0095 4583.7362 4581.6165 4579.6633 4577.8889 4576.3035
#> [2920] 4574.9150 4573.7280 4572.7439 4571.9609 4571.3751 4570.9810 4570.7729
#> [2927] 4570.7449 4570.8906 4571.2031 4571.6745 4572.2960 4573.0576 4573.9483
#> [2934] 4574.9556 4576.0656 4577.2628 4578.5310 4579.8534 4581.2134 4582.5949
#> [2941] 4583.9835 4585.3663 4586.7322 4588.0723 4589.3796 4590.6494 4591.8791
#> [2948] 4593.0678 4594.2164 4595.3270 4596.4034 4597.4511 4598.4768 4599.4892
#> [2955] 4600.4985 4601.5157 4602.5521 4603.6182 4604.7218 4605.8672 4607.0526
#> [2962] 4608.2695 4609.4998 4610.7150 4611.8734 4612.9183 4613.7757 4614.3518
#> [2969] 4614.5315 4614.1770 4613.1269 4611.1970 4608.1824 4603.8604 4597.9946
#> [2976] 4590.3385 4580.6392 4568.6415 4554.0917 4536.7427 4516.3594 4492.7250
#> [2983] 4465.6467 4434.9620 4400.5461 4362.3180 4320.2482 4274.3658 4224.7646
#> [2990] 4171.6088 4115.1372 4055.6672 3993.5977 3929.4097 3863.6659 3797.0071
#> [2997] 3730.1462 3663.8592 3598.9729 3536.3507 3476.8742 3421.4217 3370.8431
#> [3004] 3325.9321 3287.3990 3255.8491 3231.7680 3215.5167 3207.3325